Feature/package name (#12)
* Rename messed package name to edu.illinois.covid as the bundle id

* Change safer illinois logo

* Dont disable button based on the repost flag
