+++ android/app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
           xmlns:tools="http://schemas.android.com/tools"
-          package="edu.illinois.rokwire">
+          package="edu.illinois.covid">
 
     <!-- io.flutter.app.FlutterApplication is an android.app.Application that
          calls FlutterMain.startInitialization(this); in its onCreate method.
@@ -48,9 +48,9 @@
     <uses-feature android:name="android.hardware.bluetooth_le" android:required="true" />
 
     <application
-        android:name=".App"
+        android:name="edu.illinois.covid.App"
         android:allowBackup="true"
-        android:backupAgent="edu.illinois.rokwire.AppBackupAgent"
+        android:backupAgent="edu.illinois.covid.AppBackupAgent"
         android:restoreAnyVersion="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -64,7 +64,7 @@
             android:name="com.google.android.backup.api_key"
             android:value="${backupApiKey}" />
         <activity
-            android:name=".MainActivity"
+            android:name="edu.illinois.covid.MainActivity"
             android:launchMode="singleTop"
             android:theme="@style/LaunchTheme"
             android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection|fontScale|screenLayout|density|uiMode"
@@ -102,28 +102,28 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".maps.MapPickLocationActivity"
+            android:name="edu.illinois.covid.maps.MapPickLocationActivity"
             android:theme="@style/DefaultTheme"
             android:screenOrientation="portrait"
             android:configChanges="orientation|keyboardHidden"
 
         />
         <activity
-            android:name=".maps.MapDirectionsActivity"
+            android:name="edu.illinois.covid.maps.MapDirectionsActivity"
             android:theme="@style/DefaultTheme"
             android:screenOrientation="portrait"
             android:configChanges="orientation|keyboardHidden"/>
 
         <activity
-            android:name=".maps.MapActivity"
+            android:name="edu.illinois.covid.maps.MapActivity"
             android:theme="@style/DefaultTheme"
             android:screenOrientation="portrait"
             android:configChanges="orientation|keyboardHidden"/>
 
-        <service android:name=".exposure.ble.ExposureServer"/>
-        <service android:name=".exposure.ble.ExposureClient"/>
+        <service android:name="edu.illinois.covid.exposure.ble.ExposureServer"/>
+        <service android:name="edu.illinois.covid.exposure.ble.ExposureClient"/>
 
-        <receiver android:name=".exposure.ble.scan.ExposureBleReceiver">
+        <receiver android:name="edu.illinois.covid.exposure.ble.scan.ExposureBleReceiver">
             <intent-filter>
                 <action android:name="edu.illinois.rokwire.exposure.ble.scan.ACTION_FOUND" />
             </intent-filter>
