+++ android/app/src/main/java/edu/illinois/covid/exposure/crypto/AES.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package edu.illinois.rokwire.exposure.crypto;
+package edu.illinois.covid.exposure.crypto;
 
 import javax.crypto.Cipher;
 import javax.crypto.spec.SecretKeySpec;
