+++ android/app/src/main/java/edu/illinois/covid/exposure/crypto/AES_CTR.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package edu.illinois.rokwire.exposure.crypto;
+package edu.illinois.covid.exposure.crypto;
 
 import javax.crypto.Cipher;
 import javax.crypto.spec.IvParameterSpec;
