+++ android/app/src/main/java/edu/illinois/covid/maps/MapDirectionsActivity.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package edu.illinois.rokwire.maps;
+package edu.illinois.covid.maps;
 
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -66,10 +66,10 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 
-import edu.illinois.rokwire.Constants;
-import edu.illinois.rokwire.MainActivity;
-import edu.illinois.rokwire.R;
-import edu.illinois.rokwire.Utils;
+import edu.illinois.covid.Constants;
+import edu.illinois.covid.MainActivity;
+import edu.illinois.covid.R;
+import edu.illinois.covid.Utils;
 
 public class MapDirectionsActivity extends MapActivity implements OnRouteResultListener, OnLegSelectedListener {
