+++ android/app/src/main/java/edu/illinois/covid/maps/MapPickLocationActivity.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package edu.illinois.rokwire.maps;
+package edu.illinois.covid.maps;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -46,9 +46,9 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 
-import edu.illinois.rokwire.Constants;
-import edu.illinois.rokwire.R;
-import edu.illinois.rokwire.Utils;
+import edu.illinois.covid.Constants;
+import edu.illinois.covid.R;
+import edu.illinois.covid.Utils;
 
 public class MapPickLocationActivity extends AppCompatActivity {
